// Types/listingType.ts
export interface ListingType {
    id: string;
    name: string;
    image: string;
    description: string;
    rating: number;
    price: string;
    duration: string;
    location: string;
    category: string;
    Day1?: string;
    Day2?: string;
    Day3?: string;
    Day4?: string;
    Day5?: string;
    Day6?: string;
    Day7?: string;
    }
    